Hack: Add switch to only import CSS when in browser.
This prevents compile errors when spec JS is parsed on server.  Would
be nice to figure out a better way to handle this in the future.

JS is parsed on the server so that a copy of the specs are available in
memory.  This is to allow for `it.server` (yet to be implemented).